css/css-scroll-snap/parsing/scroll-padding-inline-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-padding-inline with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-padding-inline">
<meta name="assert" content="scroll-padding-inline supports the full grammar '[ auto | <length-percentage> ]{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-padding-inline", "auto");

test_valid_value("scroll-padding-inline", "10px");
test_valid_value("scroll-padding-inline", "20%");
test_valid_value("scroll-padding-inline", "calc(2em + 3ex)");
test_valid_value("scroll-padding-inline", "calc(50% + 60px)");

test_valid_value("scroll-padding-inline", "1px 2px");
test_valid_value("scroll-padding-inline", "1px auto");
test_valid_value("scroll-padding-inline", "auto auto");
</script>
</body>
</html>
